CHANGE - loading emojinote with loading animation - refactor component/App and component/GetEmojinote - remove useless fetch on app start
